feat: update container build workflow and add new message handling API integration
