+++ naraninyeo/entrypoints/http.py
@@ -1,15 +1,32 @@
 from typing import Annotated
 
+import httpx
 import uvicorn
 from fastapi import Depends, FastAPI
 
 from naraninyeo.di import container
 from naraninyeo.domain.application.new_message_handler import NewMessageHandler
 from naraninyeo.domain.model.message import Message
-from naraninyeo.entrypoints.kafka import APIClient
 from naraninyeo.infrastructure.settings import Settings
 
 app = FastAPI()
+
+
+class APIClient:
+    def __init__(self, settings: Settings):
+        self.api_url = settings.NARANINYEO_API_URL
+
+    async def send_response(self, message: Message):
+        async with httpx.AsyncClient() as client:
+            response = await client.post(
+                f"{self.api_url}/reply",
+                json={
+                    "type": "text",
+                    "room": message.channel.channel_id,
+                    "data": message.content.text,
+                },
+            )
+            response.raise_for_status()
 
 
 async def get_message_handler() -> NewMessageHandler:
@@ -26,7 +43,7 @@ async def read_root():
     return {"Hello": "World"}
 
 
-@app.get("/handle_new_message")
+@app.post("/handle_new_message")
 async def handle_new_message(
     new_message: Message,
     message_handler: Annotated[NewMessageHandler, Depends(get_message_handler)],
